Update and rename strand.rb to dna_strand.rb
